feat : 지현 - 휴가 캘린더 AOP 걸음

+++ Yolo/src/main/java/com/yolo/hr/josh/service/VacationScheduleService.java
@@ -40,7 +40,14 @@ public class VacationScheduleService implements InterVacationScheduleService {
 	@Override
 	public void delSchedule(Map<String, String> paraMap) {
 		
-		scheduledao.deleteVactionSchedule(paraMap);
+		String approval = paraMap.get("approval");
+		
+		// 반려일때만 삭제해주면 된다
+		if("2".equals(approval)) {
+			scheduledao.deleteVactionSchedule(paraMap);
+		}
+		
+		
 		
 	}
